chore: zoomable gridView, get rid of decoratorView
