+++ packages/core/src/view/unitView.js
@@ -40,7 +40,6 @@ export const markTypes = {
  * @typedef {import("../encoder/accessor").Accessor} Accessor
  * @typedef {import("../utils/layout/flexLayout").SizeDef} SizeDef
  * @typedef {import("../spec/view").ResolutionTarget} ResolutionTarget
- * @typedef {import("./decoratorView").default} DecoratorView
  *
  */
 export default class UnitView extends ContainerView {
@@ -64,7 +63,7 @@ export default class UnitView extends ContainerView {
             throw new Error(`No such mark: ${this.getMarkType()}`);
         }
 
-        /** @type {(UnitView | LayerView | DecoratorView)[]} */
+        /** @type {(UnitView | LayerView)[]} */
         this.sampleAggregateViews = [];
         this._initializeAggregateViews();
 
@@ -362,6 +361,13 @@ export default class UnitView extends ContainerView {
                 this.sampleAggregateViews.push(summaryView);
             }
         }
+    }
+
+    /**
+     * @param {import("../utils/interactionEvent").default} event
+     */
+    propagateInteractionEvent(event) {
+        event.target = this;
     }
 
     /**
